requirements.txt
python-dotenv
langchain
PyPDF2
faiss-cpu
sentence_transformers
InstructorEmbedding
chromadb
tiktoken
torch
accelerate